refactor: update import paths to use unified types module
