+++ bciers/apps/compliance/src/app/components/compliance-summary/manage-obligation/review-compliance-summary/MonetaryPaymentsGrid.tsx
@@ -5,7 +5,7 @@ import DataGrid from "@bciers/components/datagrid/DataGrid";
 import HeaderSearchCell from "@bciers/components/datagrid/cells/HeaderSearchCell";
 import monetaryPaymentsColumns from "@/compliance/src/app/components/datagrid/models/manetary-payments/monetaryPaymentsColumns";
 import monetaryPaymentsGroupColumns from "@/compliance/src/app/components/datagrid/models/manetary-payments/monetaryPaymentsGroupColumns";
-import { PaymentsData } from "@/compliance/src/app/types/payments";
+import { PaymentsData } from "@/compliance/src/app/types";
 import AlertNote from "@bciers/components/form/components/AlertNote";
 
 interface MonetaryPaymentsProps {
