+++ bciers/apps/compliance/src/app/components/compliance-summary/manage-obligation/review-compliance-summary/fetchComplianceSummaryReviewPageData.ts
@@ -1,6 +1,6 @@
 import { getComplianceSummary } from "@/compliance/src/app/utils/getComplianceSummary";
 import { getComplianceSummaryPayments } from "@/compliance/src/app/utils/getComplianceSummaryPayments";
-import { PaymentsData } from "@/compliance/src/app/types/payments";
+import { PaymentsData } from "@/compliance/src/app/types";
 
 interface ComplianceSummaryReviewPageData {
   complianceSummary: any;
