+++ bciers/apps/compliance/src/app/types.ts
@@ -20,3 +20,17 @@ export interface BccrAccountDetailsResponse {
   tradingName: string | null;
   error?: string;
 }
+
+export interface Payment {
+  id: string | number;
+  paymentReceivedDate: string;
+  paymentAmountApplied: number;
+  paymentMethod: string;
+  transactionType: string;
+  receiptNumber: string;
+}
+
+export interface PaymentsData {
+  rows: Payment[];
+  row_count: number;
+}
