--- bciers/apps/compliance/src/app/types/payments.ts
@@ -1,13 +1,0 @@
-export interface Payment {
-  id: string | number;
-  paymentReceivedDate: string;
-  paymentAmountApplied: number;
-  paymentMethod: string;
-  transactionType: string;
-  receiptNumber: string;
-}
-
-export interface PaymentsData {
-  rows: Payment[];
-  row_count: number;
-}
